[refactor] Delete 'user' packages
